Add files via upload
Found an encrypted message in a video game, originally thought it was a simple letter shift. It wasn't. But I still wrote this because, why not? Allows user to input a message and number to use for shifting the letters. Can also be used to decrypt "Caeser Cipher" messages, by entering the opposite of the shift number.

https://en.wikipedia.org/wiki/Caesar_cipher